+++ lib/ChargeBee/Models/InvoiceEstimateLineItem.php
@@ -23,6 +23,7 @@ class InvoiceEstimateLineItem extends Model
     'amountInDecimal',
     'discountAmount',
     'itemLevelDiscountAmount',
+    'referenceLineItemId',
     'description',
     'entityDescription',
     'entityType',