+++ lib/ChargeBee/Models/InvoiceShippingAddress.php
@@ -21,6 +21,7 @@ class InvoiceShippingAddress extends Model
     'country',
     'zip',
     'validationStatus',
+    'index',
   ];
 
 }