+++ lib/ChargeBee/Models/QuoteShippingAddress.php
@@ -21,6 +21,7 @@ class QuoteShippingAddress extends Model
     'country',
     'zip',
     'validationStatus',
+    'index',
   ];
 
 }