+++ lib/ChargeBee/Models/SubscriptionEstimateShippingAddress.php
@@ -21,6 +21,7 @@ class SubscriptionEstimateShippingAddress extends Model
     'country',
     'zip',
     'validationStatus',
+    'index',
   ];
 
 }